requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.129.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
h11==0.16.0
idna==3.11
psycopg==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
six==1.17.0
sqlalchemy==2.0.46
starlette==0.52.1
starlette-graphene==0.4.0
starlette-graphene3==0.6.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
neo4j>=5.0.0
python-dotenv>=1.0.0
# Optional: For better output formatting
rich>=13.0.0


